+++ packages/integrity/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.5.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @meshql/http@0.6.0
+
 ## 0.5.4
 
 ### Patch Changes